general commit

+++ fal/runsynthe.py
@@ -259,7 +259,8 @@ class Synthe(object):
     def _reset(self,fpath='./fortfile/'):
         fortlist = glob.glob('./fort.*') + glob.glob('./ROT*')
         for ff in fortlist:
-            os.remove(ff)
+            if os.path.exists(ff):
+                os.remove(ff)
 
         self._makesym('{}/fort.12'.format(fpath),'./fort.12')
         self._makesym('{}/fort.14'.format(fpath),'./fort.14')
